Initial Commit
This is a revision of the Mass Robotics Interop Standards MQTT publisher and subscriber project. The main changes are adding another level of nesting to the statusReport topics so it is nested under its UUID. This allows us to have multiple topics per different devices without needed to parse lots of unnecessary messages for other devices that also have a statusReport. the # symbol is a wildcard allowed by mosquito client for Subscribers only. Use this carefully!